video: tegra: host: extend syncpt field
This patch extends syncpoint field to 32bits to support
wider range of syncpoint ids.

Bug 1685200

Reviewed-on: http://git-master/r/797920
(cherry picked from commit eb43e2e77eabb520fb98eb36183e75d792eeccb7)
Reviewed-on: http://git-master/r/804460
GVS: Gerrit_Virtual_Submit
Tested-by: Arto Merilainen <amerilainen@nvidia.com>

+++ drivers/video/tegra/host/nvhost_intr.h
@@ -79,7 +79,7 @@ struct nvhost_intr;
 
 struct nvhost_intr_syncpt {
 	struct nvhost_intr *intr;
-	u8 id;
+	u32 id;
 	struct mutex lock;
 	struct list_head wait_head;
 	char thresh_irq_name[12];
